+++ any_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/fluentassert/verify"
 )
 
-func TestObj(t *testing.T) {
+func TestAny(t *testing.T) {
 	type A struct {
 		Str   string
 		Bool  bool
@@ -17,18 +17,18 @@ func TestObj(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
-			msg := verify.Obj(got).DeepEqual(want)
+			msg := verify.Any(got).DeepEqual(want)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "wrong", Bool: true, Slice: []int{1, 3}}
-			msg := verify.Obj(got).DeepEqual(want)
+			msg := verify.Any(got).DeepEqual(want)
 			assertFailed(t, msg, "mismatch (-want +got):\n")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := verify.Obj(got).DeepEqual(nil)
+			msg := verify.Any(got).DeepEqual(nil)
 			assertPassed(t, msg)
 		})
 	})
@@ -37,18 +37,18 @@ func TestObj(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "wrong", Bool: true, Slice: []int{1, 3}}
-			msg := verify.Obj(got).NotDeepEqual(want)
+			msg := verify.Any(got).NotDeepEqual(want)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			want := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
 			got := A{Str: "string", Bool: true, Slice: []int{1, 2, 3}}
-			msg := verify.Obj(got).NotDeepEqual(want)
+			msg := verify.Any(got).NotDeepEqual(want)
 			assertFailed(t, msg, "the objects are equal")
 		})
 		t.Run("nil", func(t *testing.T) {
 			var got *A
-			msg := verify.Obj(got).NotDeepEqual(nil)
+			msg := verify.Any(got).NotDeepEqual(nil)
 			assertFailed(t, msg, "the objects are equal")
 		})
 	})
@@ -58,14 +58,14 @@ func TestObj(t *testing.T) {
 			fn := func(x A) verify.FailureMessage {
 				return ""
 			}
-			msg := verify.Obj(A{}).Check(fn)
+			msg := verify.Any(A{}).Check(fn)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			fn := func(x A) verify.FailureMessage {
 				return "failure"
 			}
-			msg := verify.Obj(A{}).Check(fn)
+			msg := verify.Any(A{}).Check(fn)
 			assertFailed(t, msg, "failure")
 		})
 	})
@@ -75,14 +75,14 @@ func TestObj(t *testing.T) {
 			pred := func(x A) bool {
 				return true
 			}
-			msg := verify.Obj(A{}).Should(pred)
+			msg := verify.Any(A{}).Should(pred)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			pred := func(x A) bool {
 				return false
 			}
-			msg := verify.Obj(A{}).Should(pred)
+			msg := verify.Any(A{}).Should(pred)
 			assertFailed(t, msg, "object does not meet the predicate criteria")
 		})
 	})
@@ -92,14 +92,14 @@ func TestObj(t *testing.T) {
 			pred := func(x A) bool {
 				return false
 			}
-			msg := verify.Obj(A{}).ShouldNot(pred)
+			msg := verify.Any(A{}).ShouldNot(pred)
 			assertPassed(t, msg)
 		})
 		t.Run("Failed", func(t *testing.T) {
 			pred := func(x A) bool {
 				return true
 			}
-			msg := verify.Obj(A{}).ShouldNot(pred)
+			msg := verify.Any(A{}).ShouldNot(pred)
 			assertFailed(t, msg, "object meets the predicate criteria")
 		})
 	})