+++ ordered.go
@@ -15,7 +15,7 @@ type FluentOrdered[T constraints.Ordered] struct {
 // Ordered is used for testing a ordered object
 // that supports the operators < <= >= >.
 func Ordered[T constraints.Ordered](got T) FluentOrdered[T] {
-	return FluentOrdered[T]{FluentComparable[T]{FluentObj[T]{got}}}
+	return FluentOrdered[T]{FluentComparable[T]{FluentAny[T]{got}}}
 }
 
 // Lesser tests the objects using < operator.