+++ CHANGELOG.md
@@ -67,6 +67,7 @@ The main additions are the new assertions for
 ### Changed
 
 - The `f` package is renamed to `verify`.
+- Rename `Obj` and `FluentObj` to `Any` and `FluentAny`.
 - Change the `Check` assertion for `any` object so that the
   provided function has to return `FailureMessage`
   instead of a `string`.